+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Cinecred Changelog
 - Comprehensively overhauled the new project template.
 - Reduced the memory consumption of the still preview, video preview, and
   renderers, thereby allowing for pages of effectively arbitrary length.
+- Added ProRes 4444 (with optional alpha) and DNxHR 444 delivery options.
 - Unified project opening, project creation, preferences, and update
   notification in one integrated welcome window, thereby alleviating the need
   for various popup dialogs.